(refactor) extract HIW step data to data/ barrel, no behavior change, refs #38 #101
Split steps.tsx into a typed data/ module ahead of the v8 redesign so
later commits can layer contract-accurate DealState + ResolutionType
mappings without churning a single file. Pure refactor — zero runtime
diff, all 130 unit + a11y tests still pass, no SCSS or SVG touched.

- data/types.ts         — HiwStep, HiwChoreo, LedgerState, ActorId,
                          WireId, PhaseIndex, PHASE_COUNT
- data/geometry.ts      — SVG_VIEW_BOX, ACTOR_POSITIONS, CORE_POSITION
- data/phases.tsx       — HIW_STEPS (JSX narration preserved verbatim)
- data/ledger-logs.tsx  — LEDGER_LOGS + LedgerLog type
- data/index.ts         — barrel

Importers updated: HowItWorks.tsx, HowItWorks.test.tsx,
HowItWorksAnimations.tsx, HowItWorksAnimations.test.tsx.
