+++ apps/web/src/features/homepage/HowItWorks/data/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './geometry';
+export * from './phases';
+export * from './ledger-logs';
